Major refactor of the teams
