--- src/components/views/SecurityTeams/style.scss
@@ -1,60 +1,0 @@
-.security-teams {
-  textarea {
-    resize: none;
-  }
-  .card {
-    max-height: 60vh;
-    overflow-y: scroll;
-    p {
-      cursor: pointer;
-      margin: 0;
-      padding: 5px;
-      border-radius: 5px;
-      &:active {
-        background-color: rgba(255, 255, 255, 0.4);
-      }
-      &.selected {
-        background-color: rgba(255, 255, 255, 0.2);
-      }
-    }
-  }
-  .officers {
-    .card {
-      height: 25vh;
-      overflow-y: scroll;
-    }
-  }
-  .location-label,
-  .location-label label {
-    margin-bottom: 0;
-    padding-bottom: 0;
-  }
-  .row .recall-button {
-    margin-top: 50px;
-  }
-  .form-control-lg {
-    font-size: 14px;
-  }
-}
-.security-teams-core {
-  height: 100%;
-  .row {
-    height: 100%;
-  }
-  label {
-    display: block;
-    margin: 0;
-    font-weight: 800;
-  }
-  .scroller {
-    max-height: calc(100% - 20px);
-    overflow-y: scroll;
-  }
-  .label-section {
-    margin-bottom: 5px;
-  }
-  p {
-    margin: 0;
-    margin-bottom: 5px;
-  }
-}
