+++ src/components/views/Teams/security.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Teams from "./index";
+
+const SecurityTeams = props => {
+  return <Teams {...props} teamType="security" />;
+};
+
+export default SecurityTeams;
