+++ src/components/views/Teams/securityCore.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Core from "./core";
+
+const SecurityTeamsCore = props => {
+  return <Core {...props} teamType="security" />;
+};
+
+export default SecurityTeamsCore;
