+++ src/components/views/Teams/training.js
@@ -1,0 +1,120 @@
+export default {
+  security: [
+    {
+      selector: ".nothing",
+      content:
+        "Security officers are other people on the ship assigned to keep the entire crew safe. You are responsible for understanding what threats are on the ship and assigning security officers to take care of those threats."
+    },
+    {
+      selector: ".team-list",
+      content:
+        "This is the list of all assigned security teams. If you are just getting started, it should be empty."
+    },
+    {
+      selector: ".new-button",
+      content: "Click this button to create a new team."
+    },
+    {
+      selector: ".team-name",
+      content:
+        "Before the team can be created, we need to fill in some information. Start by naming your security team. This is what you will use to identify the security team later on."
+    },
+    {
+      selector: ".team-orders",
+      content:
+        "Give your team some orders. This is what they are supposed to do. It could be something like investigating, patroling, taking care of intruders, or guarding something. Your team will obey your orders."
+    },
+    {
+      selector: ".team-location",
+      content:
+        "You also have to tell your security team members where to go. You first need to pick a deck. Once you have picked a deck, you can optionally select a room for them to go to. This helps them know exactly where you want them to be."
+    },
+    {
+      selector: ".crew-list",
+      content:
+        "You have to assign officers to be on the team. Otherwise, it will be empty and nothing will happen. Click on the names above to assign those officers to your team. You have a limited number of officers, and they can only be assigned to one team at a time, so use them carefully."
+    },
+    {
+      selector: ".crew-assigned",
+      content:
+        "This is where your assigned officers show up. If you want to remove an officer from the team, click on the officer's name."
+    },
+    {
+      selector: ".create-button",
+      content:
+        "Any time you create a team, or whenever you make changes, you have to click the 'Create' or 'Update' button here."
+    },
+    {
+      selector: ".cancel-button",
+      content:
+        "If you change your mind, or want to recall your officers and remove the team, you can click on the 'Cancel' or 'Recall' button here."
+    },
+    {
+      selector: "#widget-messages",
+      content:
+        "You can communicate with your security teams using the messaging widget. You can contact the teams individually, or you can communicate with the security station. Be sure you regularly check your messages and respond promptly."
+    }
+  ],
+  damage: [
+    {
+      selector: ".nothing",
+      content:
+        "There are officers on the ship assigned to repair and maintain the ship. You are responsible for keeping track of damage and maintenance and assigning these officers to handle those problems."
+    },
+    {
+      selector: ".team-list",
+      content:
+        "This is the list of all assigned damage teams. If you are just getting started, it should be empty."
+    },
+    {
+      selector: ".new-team",
+      content: "Click this button to create a new team."
+    },
+    {
+      selector: ".team-name",
+      content:
+        "Before the team can be created, we need to fill in some information. Start by naming your damage team. This is what you will use to identify the team later on."
+    },
+    {
+      selector: ".team-orders",
+      content:
+        "Give your team some orders. This is what they are supposed to do. It could be something repairing a system, running diagnostics, or cleaning up. Your damage reports should tell you what orders to give the teams."
+    },
+    {
+      selector: ".team-location",
+      content:
+        "You also have to tell your damage team members where to go. You first need to pick a deck. Once you have picked a deck, you can optionally select a room for them to go to. This helps them know exactly where you want them to be."
+    },
+    {
+      selector: ".team-priority",
+      content:
+        "Some work orders are more important than others. Assign a priority to this team with these buttons."
+    },
+    {
+      selector: ".crew-list",
+      content:
+        "You have to assign officers to be on the team. Each officer has a specific thing that they are assigned to do, so make sure you choose the right person for the job. Click on the names above to assign those officers to your team. You have a limited number of officers, and they can only be assigned to one team at a time, so use them carefully."
+    },
+    {
+      selector: ".crew-assigned",
+      content:
+        "This is where your assigned officers show up. If you want to remove an officer from the team, click on the officer's name."
+    },
+    {
+      selector: ".create-button",
+      content:
+        "Any time you create a team, or whenever you make changes, you have to click the 'Create' or 'Update' button here."
+    },
+    {
+      selector: ".cancel-button",
+      content:
+        "If you change your mind, or want to recall your officers and remove the team, you can click on the 'Cancel' or 'Recall' button here."
+    },
+    {
+      selector: "#widget-messages",
+      content:
+        "You can communicate with your damage teams using the messaging widget. You can contact the teams individually, or you can communicate with the damage team station. Be sure you regularly check your messages and respond promptly."
+    }
+  ],
+  medical: []
+};
