+++ src/components/views/index.js
@@ -24,8 +24,8 @@ import CargoControl from "./CargoControl";
 import CoolantControl from "./CoolantControl";
 import TractorBeam from "./TractorBeam";
 
-import SecurityTeams from "./SecurityTeams";
-import DamageTeams from "./DamageTeams";
+import SecurityTeams from "./Teams/security";
+import DamageTeams from "./Teams";
 import AlertCondition from "./AlertCondition";
 import ProbeNetwork from "./ProbeNetwork";
 import SelfDestruct from "./SelfDestruct";
@@ -81,8 +81,8 @@ import ActionsCore from "./Actions/core";
 import ThrusterCore from "./Thrusters/core";
 import TractorBeamCore from "./TractorBeam/core";
 import CrewCore from "./Crew/core";
-import SecurityTeamsCore from "./SecurityTeams/core";
-import DamageTeamsCore from "./DamageTeams/core";
+import SecurityTeamsCore from "./Teams/securityCore";
+import DamageTeamsCore from "./Teams/core";
 import ProbeNetworkCore from "./ProbeNetwork/core";
 import SelfDestructCore from "./SelfDestruct/core";
 import ProbeControlCore from "./ProbeControl/core";
